+++ hydra-node/src/Hydra/API/WSServer.hs
@@ -44,6 +44,7 @@ import Hydra.HeadLogic (ClosedState (ClosedState, readyToFanoutSent), HeadState,
 import Hydra.HeadLogic.State qualified as HeadState
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.NetworkVersions qualified as NetworkVersions
+import Hydra.Node.Environment (Environment (..))
 import Hydra.Tx (Party)
 import Network.WebSockets (
   PendingConnection (pendingRequest),
@@ -59,6 +60,7 @@ import Text.URI.QQ (queryKey, queryValue)
 wsApp ::
   forall tx.
   IsChainState tx =>
+  Environment ->
   Party ->
   Tracer IO APIServerLog ->
   ConduitT () (TimedServerOutput tx) (ResourceT IO) () ->
@@ -69,7 +71,7 @@ wsApp ::
   ServerOutputFilter tx ->
   PendingConnection ->
   IO ()
-wsApp party tracer history callback headStateP responseChannel ServerOutputFilter{txContainsAddr} pending = do
+wsApp env party tracer history callback headStateP responseChannel ServerOutputFilter{txContainsAddr} pending = do
   traceWith tracer NewAPIConnection
   let path = requestPath $ pendingRequest pending
   queryParams <- uriQuery <$> mkURIBs path
@@ -101,6 +103,7 @@ wsApp party tracer history callback headStateP responseChannel ServerOutputFilte
             , hydraHeadId = getHeadId headState
             , snapshotUtxo = getSnapshotUtxo headState
             , hydraNodeVersion = showVersion NetworkVersions.hydraNodeVersion
+            , env
             }
 
   Projection{getLatest} = headStateP